FreezeCurves

FreezeCurves.BrooksCorey — Type

BrooksCorey{Twp,Tψₛ,Tλ} <: SWRC

van Genuchten MT, 1980. A closed-form equation for predicting the hydraulic conductivity of unsaturated soils. Soil Science Society of America Journal, 44(5): 892–898. DOI: 10.2136/sssaj 1980.03615995004400050002x.

FreezeCurves.DallAmico — Type

DallAmico{TFT,Tg,Tswrc<:SWRC} <: SFCC

Dall'Amico M, 2010. Coupled water and heat transfer in permafrost modeling. Ph.D. Thesis, University of Trento, pp. 43.

FreezeCurves.DallAmicoSalt — Type

DallAmicoSalt{TFT,Tsc,TR,Tg,Tswrc<:SWRC} <: SFCC

Freeze curve from Dall'Amico (2011) with saline freezing point depression.

Angelopoulos M, Westermann S, Overduin P, Faguet A, Olenchenko V, Grosse G, Grigoriev MN. Heat and salt flow in subsea permafrost modeled with CryoGRID2. Journal of Geophysical Research: Earth Surface. 2019 Apr;124(4):920-37.

FreezeCurves.FreeWater — Type

FreeWater <: FreezeCurve

"Free water" freeze curve in terms of enthalpy (H), total water content (θtot), and the latent heat of fusion of water (L).

FreezeCurves.FreezeCurve — Type

FreezeCurve

Base type for freezing characteristic curves which relate temperature or enthalpy to unfrozen water content.

FreezeCurves.Hu2020 — Type

Hu2020{TFT,Tvol,Tb}  <: SFCC

Soil freezing characteristic curve formulation of Hu et al. 2020.

Hu G, Zhao L, Zhu X, Wu X, Wu T, Li R, Xie C, Hao J. Review of algorithms and parameterizations to determine unfrozen water content in frozen soil. Geoderma. 2020 Jun 1;368:114277.

FreezeCurves.McKenzie — Type

McKenzie{TFT,Tvol,Tγ} <: SFCC

McKenzie JM, Voss CI, Siegel DI, 2007. Groundwater flow with energy transport and water-ice phase change: numerical simulations, benchmarks, and application to freezing in peat bogs. Advances in Water Resources, 30(4): 966–983. DOI: 10.1016/j.advwatres.2006.08.008.

FreezeCurves.PainterKarra — Type

PainterKarra{TFT,Tβ,Tω,Tg,Tswrc<:SWRC} <: SFCC

Painter SL, Karra S. Constitutive Model for Unfrozen Water Content in Subfreezing Unsaturated Soils. Vadose zone j. 2014 Apr;13(4):1-8.

FreezeCurves.PowerLaw — Type

PowerLaw{Tvol,Tα,Tβ} <: SFCC

Commonly used power law relation of Lovell (1957).

Lovell, C.: Temperature effects on phase composition and strength of partially frozen soil, Highway Research Board Bulletin, 168, 74–95, 1957.

FreezeCurves.SFCC — Type

SFCC <: FreezeCurve

Base type for a soil freeze characteristic curve (SFCC) function. Subtypes should be callable structs that implement the freeze curve and contain any necessary additional constants or configuration options. User-specified parameters can either be supplied in the struct or declared as model parameters via the variables method.

FreezeCurves.SFCCSolver — Type

SFCCSolver

Base type representing non-linear solvers for implicit SFCC functions.

FreezeCurves.SWRC — Type

SWRC

Base type for soil water retention curves (SWRC) which relate soil water matric potential to water content.

FreezeCurves.SoilFreezeThawProperties — Type

SoilFreezeThawProperties{TTₘ,TLsl}

Struct containing constants/parameters common to some or all SFCCs.

FreezeCurves.SoilFreezeThawProperties — Method

SoilFreezeThawProperties(f::SFCC)

Retrieves the default SoilFreezeThawProperties from f; should be defined for all freeze curves.

FreezeCurves.SoilWaterVolume — Type

SoilWaterVolume{Tρw,Tθres,Tθsat}

Struct containing basic physical constants and propeties related to soil pore water.

FreezeCurves.SoilWaterVolume — Method

SoilWaterVolume(f::SFCC)

Retrieves the nested SoilWaterVolume from the SoilFreezeThawProperties of the freeze curve f.

FreezeCurves.SoilWaterVolume — Method

SoilWaterVolume(f::SWRC)

Get the SoilWaterVolume defined for the given SWRC f. Must be implemented for all SWRC types. Default implementation retrieves the field water.

FreezeCurves.VanGenuchten — Type

VanGenuchten{Tvol,Tα,Tn} <: SWRC

van Genuchten MT, 1980. A closed-form equation for predicting the hydraulic conductivity of unsaturated soils. Soil Science Society of America Journal, 44(5): 892–898. DOI: 10.2136/sssaj 1980.03615995004400050002x.

FreezeCurves.Westermann — Type

Westermann{TFT,Tvol,Tδ} <: SFCC

Westermann, S., Boike, J., Langer, M., Schuler, T. V., and Etzelmüller, B.: Modeling the impact of wintertime rain events on the thermal regime of permafrost, The Cryosphere, 5, 945–959, https://doi.org/10.5194/tc-5-945-2011, 2011.

FreezeCurves.inflectionpoint — Method

inflectionpoint(f::SFCC)

Returns the analytical inflection point (i.e. where ∂²θ/∂T^2 = 0), if available.

FreezeCurves.inflectionpoint — Method

inflectionpoint(f::SWRC)

Returns the analytical solution for the inflection point (i.e. where ∂²θ/∂ψ² = 0) of the SWRC, if available.

Unitful.ustrip — Method

ustrip(x)

Reconstructs the type or function x with all numerical quantities stripped of units.