Solvers

The Solvers sub-module provides an interface for performing non-linear root-finding, e.g. solving implicit freeze curve formulations or recovering temperature given enthalpy, on the soil freeze characteristic curves implemented in this package. Note that this is not to be confused with parameter inference which is handled by a separate sub-module.

FreezeCurves.Solvers.SFCCInverseEnthalpyObjective — Type

SFCCInverseEnthalpyObjective{TF,Tkwargs<:NamedTuple,Thc,TL,TH} <: AbstractSFCCObjective

Optimization objective for finding a temperature, T, wich resolves the conservation equation:

\[0 = (H - Lθ)/C - T\]

given a fixed enthalpy value H and freeze curve arguments f_args.

FreezeCurves.Solvers.SFCCNewtonSolver — Type

SFCCNewtonSolver <: SFCCSolver

Fast, specialized implementation of Newton's method with backtracking line search for resolving the energy conservation law, H = TC + Lθ. Attempts to find the root of the corresponding temperature residual: ϵ = T - (H - Lθ(T)) / C(θ(T)) and uses backtracking to avoid jumping over the solution. This prevents convergence issues that arise due to discontinuities and strong non-linearity in most common soil freeze curves.

FreezeCurves.Solvers.SFCCPreSolver — Type

SFCCPreSolver{TCache} <: SFCCSolver

A fast SFCC "solver" which pre-builds an interpolant for the freeze curve in terms of enthalpy, θ(H). Note that this solver is only valid when all freeze curve parameters are held constant and will produce incorrect results otherwise.

FreezeCurves.Solvers.heatcapacity — Method

heatcapacity(c_frozen::Number, c_thawed::Number; θtot=1.0)

Piecewise constant heat capacity function:

\[f(θw) = \begin{cases} c_f & θw < θtot \\ c_t & \text{otherwise} \end{cases}\]

FreezeCurves.Solvers.heatcapacity — Method

heatcapacity(c::Number)

Heat capacity function f(H,T,θw) = c where c is a pre-specified constant.

FreezeCurves.Solvers.sfccsolve — Method

sfccsolve(obj::AbstractSFCCObjective, solver::SFCCSolver, x₀, ::Val{return_all}=Val{true}()) where {return_all}

Solve the given objective obj using solver and initial guess x₀. If return_all=true, then sfccsolve should return a named tuple with at least the temperature solution T, the liquid water content θw, the heat capacity C, and the liquid water content deriviative dθdT defined. Solver-specific additional values may also be included.

FreezeCurves.Solvers.sfccsolve — Method

sfccsolve(obj::SFCCInverseEnthalpyObjective, solver::SFCCNewtonSolver, T₀::Number, ::Val{return_all}=Val{true}(); error_when_not_converged=true)

Solves obj using the specialized Newton solver and returns the result. If return_all is true, a named tuple (;T, Tres, θw, C, itercount) is returned; otherwise (by default), only the solution temperature is returned.