FreezeCurves

FreezeCurves.BrooksCorey — Type

BrooksCorey{Twp,Tψₛ,Tλ} <: SWRCFunction

van Genuchten MT, 1980. A closed-form equation for predicting the hydraulic conductivity of unsaturated soils. Soil Science Society of America Journal, 44(5): 892–898. DOI: 10.2136/sssaj 1980.03615995004400050002x.

FreezeCurves.DallAmico — Type

DallAmico{Tftp,Tg,Tswrc<:SWRCFunction} <: SFCCFunction

Dall'Amico M, 2010. Coupled water and heat transfer in permafrost modeling. Ph.D. Thesis, University of Trento, pp. 43.

FreezeCurves.DallAmicoSalt — Type

DallAmicoSalt{Tftp,Tsc,TR,Tg,Tswrc<:SWRCFunction} <: SFCCFunction

Freeze curve from Dall'Amico (2011) with saline freezing point depression.

Angelopoulos M, Westermann S, Overduin P, Faguet A, Olenchenko V, Grosse G, Grigoriev MN. Heat and salt flow in subsea permafrost modeled with CryoGRID2. Journal of Geophysical Research: Earth Surface. 2019 Apr;124(4):920-37.

FreezeCurves.McKenzie — Type

McKenzie{Tftp,Twp,Tγ} <: SFCCFunction

McKenzie JM, Voss CI, Siegel DI, 2007. Groundwater flow with energy transport and water-ice phase change: numerical simulations, benchmarks, and application to freezing in peat bogs. Advances in Water Resources, 30(4): 966–983. DOI: 10.1016/j.advwatres.2006.08.008.

FreezeCurves.PainterKarra — Type

PainterKarra{Tftp,Tβ,Tω,Tg,Tswrc<:SWRCFunction} <: SFCCFunction

Painter SL, Karra S. Constitutive Model for Unfrozen Water Content in Subfreezing Unsaturated Soils. Vadose zone j. 2014 Apr;13(4):1-8.

FreezeCurves.SFCC — Type

SFCC{F,S} <: FreezeCurve

Generic representation of the soil freezing characteristic curve along with a nonlinear solver for resolving the temperature-energy conservation law. The shape and parameters of the curve are determined by the implementation of SFCCFunction f.

FreezeCurves.SFCCFunction — Type

SFCCFunction

Base type for a soil freeze characteristic curve (SFCC) function. Subtypes should be callable structs that implement the freeze curve and contain any necessary additional constants or configuration options. User-specified parameters can either be supplied in the struct or declared as model parameters via the variables method.

FreezeCurves.SFCCSolver — Type

SFCCSolver

Base type representing non-linear solvers for implicit SFCC functions.

FreezeCurves.SWRCFunction — Type

SWRCFunction

Base type for soil water retention curves (SWRC) which relate soil water matric potential to water content.

FreezeCurves.SoilFreezeThawProperties — Type

SoilFreezeThawProperties{TTₘ,TLsl}

Struct containing constants/parameters common to some or all SFCCs.

FreezeCurves.SoilFreezeThawProperties — Method

SoilFreezeThawProperties(f::SFCCFunction)

Retrieves the default SoilFreezeThawProperties from f; should be defined for all freeze curves.

FreezeCurves.SoilWaterProperties — Type

SoilWaterProperties{Tρw,Tθres,Tθtot,Tθsat}

Struct containing basic physical constants and propeties related to soil pore water.

FreezeCurves.SoilWaterProperties — Method

SoilWaterProperties(f::SFCCFunction)

Retrieves the nested SoilWaterProperties from the SoilFreezeThawProperties of the freeze curve f.

FreezeCurves.SoilWaterProperties — Method

SoilWaterProperties(f::SWRCFunction)

Get the SoilWaterProperties defined for the given SWRCFunction f. Must be implemented for all SWRCFunction types. Default implementation retrieves the field water.

FreezeCurves.VanGenuchten — Type

VanGenuchten{Twp,Tα,Tn} <: SWRCFunction

van Genuchten MT, 1980. A closed-form equation for predicting the hydraulic conductivity of unsaturated soils. Soil Science Society of America Journal, 44(5): 892–898. DOI: 10.2136/sssaj 1980.03615995004400050002x.

FreezeCurves.Westermann — Type

Westermann{Tftp,Twp,Tδ} <: SFCCFunction

Westermann, S., Boike, J., Langer, M., Schuler, T. V., and Etzelmüller, B.: Modeling the impact of wintertime rain events on the thermal regime of permafrost, The Cryosphere, 5, 945–959, https://doi.org/10.5194/tc-5-945-2011, 2011.

Unitful.ustrip — Method

ustrip(x)

Reconstructs the type or function x with all numerical quantities stripped of units.